requirements.txt
Flask==3.0.3
python-dotenv==0.19.0
requests==2.26.0
SQLite3==3.36.0
pytest==6.2.5